fix(mcp): add request parameter to health_check and refactor system utils
1. Added GetHealthCheckRequest parameter to health_check tool
   - Prevents validation errors when MCP clients (like Claude Desktop) call the tool with empty objects
   - Follows same pattern as get_instance_info tool

2. Created system_utils.py module
   - Moved helper functions from get_instance_info.py to new system_utils.py
   - calculate_dashboard_breakdown
   - calculate_database_breakdown
   - calculate_instance_summary
   - calculate_recent_activity
   - calculate_popular_content
   - Follows established pattern used in chart/, dashboard/, and sql_lab/ modules

3. Updated get_instance_info.py
   - Imports helper functions from system_utils
   - Removed duplicate function definitions
   - Cleaner, more maintainable code structure

4. Added comprehensive unit tests
   - Tests for GetHealthCheckRequest schema validation
   - Tests for HealthCheckResponse schema
   - Tests verify MCP client compatibility (empty object acceptance)

All tests passing and pre-commit hooks satisfied.
